backend/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar
python-decouple==3.8
django-cors-headers==4.6.0
mysqlclient==2.2.4
gunicorn==23.0.0
